Move useEffectOnceIf to VirtualizedList folder
