+++ frontend/src/metabase/common/components/VirtualizedList/use-effect-once-if.ts
